Cland compat: hcore compiles and links.

+++ hcore/hthread.cxx
@@ -72,9 +72,10 @@ HThread::HThread( void )
 	M_ENSURE( ::pthread_attr_init( attr ) == 0 );
 	size_t stackSize( 0 );
 	M_ENSURE( pthread_attr_getstacksize( attr, &stackSize ) == 0 );
-	log( LOG_TYPE::DEBUG ) << "Default thread stack size: " << stackSize << " and we will use: " << ( _threadStackSize ? _threadStackSize : stackSize ) << endl;
+	log( LOG_TYPE::DEBUG ) << "Default thread stack size: " << stackSize
+		<< " and we will use: " << ( _threadStackSize ? _threadStackSize : static_cast<int>( stackSize ) ) << endl;
 	if ( _threadStackSize > 0 )
-		M_ENSURE( pthread_attr_setstacksize( attr, _threadStackSize ) == 0 );
+		M_ENSURE( pthread_attr_setstacksize( attr, static_cast<size_t>( _threadStackSize ) ) == 0 );
 	HResource<void> res( attr, do_pthread_attr_destroy );
 	_resGuard.swap( res );
 	M_ENSURE( ::pthread_attr_setdetachstate( attr, PTHREAD_CREATE_JOINABLE ) == 0 );
