+++ hcore/htime.cxx
@@ -273,12 +273,12 @@ HString HTime::string( void ) const {
 	M_PROLOG
 #ifdef HAVE_SMART_STRFTIME
 	static int const MIN_TIME_STRING_LENGTH( 32 );
-	int long size( ::strftime( NULL, 1024, _format.raw(), &_broken ) + 1 );
+	int long size( static_cast<int long>( ::strftime( NULL, 1024, _format.raw(), &_broken ) + 1 ) );
 	if ( size < 2 )
 		M_THROW( "bad format", errno );
 	_cache.realloc( max<int long>( size, MIN_TIME_STRING_LENGTH ) );
 	M_ENSURE( static_cast<int>( ::strftime( _cache.get<char>(),
-					size, _format.raw(), &_broken ) ) < size );
+					static_cast<size_t>( size ), _format.raw(), &_broken ) ) < size );
 #else /* HAVE_SMART_STRFTIME */
 	static int const MIN_TIME_STRING_LENGTH( 64 );
 	_cache.realloc( MIN_TIME_STRING_LENGTH ); /* FIXME that is pretty dumb hack */
